* src/print_graph.c (print_actions): Remove `rule', unused. * src/LR0.c (kernel_size): Contain `int' instead of `size_t' to pacify GCC's signed < unsigned warnings. * src/closure.c (itemsetsize): Likewise. * src/reader.c (symbol_list_new): Static.

+++ src/closure.c
@@ -28,7 +28,7 @@
 
 /* ITEMSETSIZE is the size of the array ITEMSET.  */
 short *itemset;
-size_t itemsetsize;
+int itemsetsize;
 
 static unsigned *ruleset;
